Make timeout calculation correct
Small refactoring around time things.
Fixed timeout calculation. Till now it has been calculated as if start
of all tests was same and it was begining of tests. Now we are marking
start time of each execution (time of creating Execution instance)
and use it in duration calculation per each test execution separate.